MediaSession integration moved from own custom element to <cyp-app> class
