+++ app/css/cyp.less
@@ -90,4 +90,3 @@ select {
 @import "elements/back.less";
 @import "elements/path.less";
 @import "elements/yt-result.less";
-@import "elements/media-handler.less";
