--- app/css/elements/media-handler.less
@@ -1,3 +1,0 @@
-cyp-media-handler {
-	display: none;
-}
