+++ app/cyp.js
@@ -587,6 +587,9 @@ class App extends HTMLElement {
 
 		await this._connect();
 		this.dispatchEvent(new CustomEvent("load"));
+
+		this.mediaSessionInit = false;
+		this._initMediaHandler();
 	}
 
 	attributeChangedCallback(name, oldValue, newValue) {
@@ -627,6 +630,49 @@ class App extends HTMLElement {
 			}
 		}
 		alert(`Failed to connect to MPD after ${attempts} attempts. Please reload the page to try again.`);
+	}
+
+	_initMediaHandler() {
+		// check support mediaSession
+		if (!('mediaSession' in navigator)) {
+			console.log('mediaSession is not supported');
+			return;
+		}
+
+		// DOM (using media session controls are allowed only if there is audio/video tag)
+		const audio = node("audio", {loop: true}, "", this);
+		node("source", {src: 'https://raw.githubusercontent.com/anars/blank-audio/master/10-seconds-of-silence.mp3'}, '', audio);
+
+		// Init event session (play audio) on click (because restrictions by web browsers)
+		window.addEventListener('click', () => {
+			if (!this.mediaSessionInit) {
+				audio.play();
+				this.mediaSessionInit = true;
+			}
+		});
+
+		// mediaSession define metadata
+		navigator.mediaSession.metadata = new MediaMetadata({
+			title: 'Control Your Player'
+		});
+
+		// mediaSession define action handlers
+		navigator.mediaSession.setActionHandler('play', () => {
+			this.mpd.command("play");
+			audio.play();
+		});
+		navigator.mediaSession.setActionHandler('pause', () => {
+			this.mpd.command("pause 1");
+			audio.pause();
+		});
+		navigator.mediaSession.setActionHandler('previoustrack', () => {
+			this.mpd.command("previous");
+			audio.play();
+		});
+		navigator.mediaSession.setActionHandler('nexttrack', () => {
+			this.mpd.command("next");
+			audio.play();
+		});
 	}
 }
 
@@ -1866,55 +1912,6 @@ class Library extends Component {
 }
 
 customElements.define("cyp-library", Library);
-
-class MediaHandler extends Component {
-	connectedCallback() {
-		// check support mediaSession
-		if (!('mediaSession' in navigator)) {
-			console.log('mediaSession is not supported');
-			return;
-		}
-
-		// DOM
-		const audio = node("audio", {loop: true}, "", this);
-		node("source", {src: 'https://raw.githubusercontent.com/anars/blank-audio/master/10-seconds-of-silence.mp3'}, '', audio);
-
-		// Init event session (play audio) on click (because restrictions by web browsers)
-		let mediaSessionInit = false;
-		window.addEventListener('click', () => {
-			if (!mediaSessionInit) {
-				audio.play();
-				mediaSessionInit = true;
-			}
-		});
-
-		// mediaSession define metadata
-		navigator.mediaSession.metadata = new MediaMetadata({
-			title: 'Control Your Player'
-		});
-
-		// mediaSession define action handlers
-		const that = this;
-		navigator.mediaSession.setActionHandler('play', function() {
-			that._mpd.command("play");
-			audio.play();
-		});
-		navigator.mediaSession.setActionHandler('pause', function() {
-			that._mpd.command("pause 1");
-			audio.pause();
-		});
-		navigator.mediaSession.setActionHandler('previoustrack', function() {
-			that._mpd.command("previous");
-			audio.play();
-		});
-		navigator.mediaSession.setActionHandler('nexttrack', function() {
-			that._mpd.command("next");
-			audio.play();
-		});
-	}
-}
-
-customElements.define("cyp-media-handler", MediaHandler);
 
 function updateSize() {
 	document.body.style.setProperty("--vh", window.innerHeight/100);
