+++ app/index.html
@@ -96,7 +96,6 @@
 				<button data-for="settings" data-icon="settings"><span>Settings</span></button>
 			</cyp-menu>
 		</footer>
-		<cyp-media-handler></cyp-media-handler>
 		</cyp-app>
 
 		<script type="module" src="cyp.js"></script>
